requirements.txt
ollama==0.4.7
faiss-cpu==1.10.0
langchain==0.3.20
langchain-community==0.3.19
sentence-transformers==3.4.1